feat(scale): H100/H200 support — int64 step counters, coherent batch scaler, TF32
Make large-GPU training stable and possible:

- int64 step counters: total_timesteps, RolloutCollector::global_step_,
  UpdateStats/BR Result.global_step, and the metrics/BR signatures. A
  scaled run easily exceeds 2^31 env steps (S=4 → 2.4B); int32 would
  silently wrap and corrupt step-denominated cadences + logs mid-run.

- tools/scale_config.sh <S> [--apply]: scales the BATCH by S while
  preserving the self-play stability invariants — num_envs,
  num_minibatches, total_timesteps ×S and learning_rate ×√S. This holds
  num_updates (policy iterations, the R-NaD/MMD last-iterate quantity)
  and minibatch_size constant; only per-update variance changes (lower =
  stabilising). Dynamics-coherent but still BR-validated. Does NOT grow
  minibatch_size (the SM-utilisation knob — separate, BR-gated).

- TF32 tensor-core path for fp32 matmuls (main.cpp; POKER_PPO_NO_TF32 to
  disable). bf16 autocast already covers the hot GEMMs, so this mainly
  helps eager/fallback paths on Ampere+/H100.

- config.h documents the big-GPU strategy: a 3M-param net can't saturate
  an H100 (low arithmetic intensity); spend headroom on a larger batch
  (scale_config) and/or a bigger MODEL (the real capacity lever), each
  BR-validated.

87/87 tests pass; default config unchanged (768x64, 600M).

+++ include/config.h
@@ -116,7 +116,7 @@ struct PPOConfig {
     int   update_epochs    = 4;
     int   num_minibatches  = 4;
 
-    int   total_timesteps  = 10'000'000;
+    int64_t total_timesteps = 10'000'000;  // int64: H100-scale runs exceed 2^31
 
     // Periodic full-state checkpoint cadence, in env steps (0 disables).
     // ~50 checkpoints over a 600M-step run; each is net+magnet+Adam moments
@@ -160,7 +160,9 @@ struct PPOConfig {
 
     constexpr int batch_size()     const noexcept { return num_envs * num_steps; }
     constexpr int minibatch_size() const noexcept { return batch_size() / num_minibatches; }
-    constexpr int num_updates()    const noexcept { return total_timesteps / batch_size(); }
+    constexpr int num_updates()    const noexcept {
+        return static_cast<int>(total_timesteps / batch_size());  // fits int
+    }
 };
 
 // Wraps Game::DefaultGameConfig + PPO-side knobs. `hist`/`round_summary`
@@ -273,6 +275,17 @@ static constexpr PPOConfig kPPOConfig{
     // (128→64 steps), negligible at gae_lambda=0.90 (~10-step effective
     // horizon). Use tools/bench_throughput.sh to re-find the knee on a
     // different GPU; the update half is compute-bound and unaffected by this.
+    //
+    // Big GPU (H100/H200): a 3M-param net can't saturate the tensor cores
+    // (arithmetic intensity too low) — it just runs fast at low util. The
+    // stable way to spend the headroom is a LARGER BATCH, scaled coherently
+    // with tools/scale_config.sh <S> (num_envs, num_minibatches,
+    // total_timesteps ×S; learning_rate ×√S) which preserves the policy-
+    // iteration count and minibatch size, then BR-validated. Bigger
+    // minibatch (to fill SMs) and a bigger MODEL (hidden_dim/num_layers/
+    // attn_dim — the real way to USE the capacity, likely a stronger bot)
+    // are separate dynamics changes; sweep + BR-validate each. Step
+    // counters are int64, so multi-billion-step runs are safe.
     .num_envs         = 768,
     .num_steps        = 64,
     .update_epochs    = 4,
